Create password_gen.py
This code defines a GUI for generating passwords. The GUI includes options for the length of the password and the character sets to use (lowercase letters, uppercase letters, digits, and punctuation), as well as a button to generate a password and a text field to display the generated password.

The GUI is implemented using the tkinter library in Python. The main class is `GUI`, which creates and lays out the widgets for the GUI. `OptControl` is a helper class for creating checkboxes for the character set options. The `PasswordGenerator` class encapsulates the logic for generating passwords.

When the user selects or deselects a checkbox or changes the password length, the `opt_changed` method of the `GUI` class is called, which updates the `PasswordGenerator` with the new options and length. When the user clicks the "Generate password" button, the `generate_password` method of the `GUI` class is called, which generates a new password using the `PasswordGenerator` and displays it in the text field.

Overall, the code is well-organized and uses good object-oriented design principles. The use of data classes for the `PasswordGenerator` and `OptControl` classes is a nice touch that helps to encapsulate their data and behavior.